ci: add GitHub workflow to lint commit messages #4
This commit adds a workflow that uses `git-cliff` to lint the commits in
the repo on every PR.

There is a complementary `cliff.toml` configuration file which is used to
configure `git-cliff` within the workflow.

I've also added extension recommendations for GitHub Actions and TOML
files to the VSCode workspace.

Fixes #4.