ci: add workflow to build and publish the document #5
This commit adds the following workflows:
- `.github/workflows/build-latex.yml`
- `.github/workflows/build-release.yml`

`.github/workflows/build-latex.yml` is a re-usable workflow that can be
used to build the LaTeX document into a PDF, and upload it as an
artifact.

`.github/workflows/build-release.yml` is a release workflow that uses
`.github/workflows/build-latex.yml` to build the LaTeX document,
generate a changelog based on the commits, and publish them to a release
on GitHub.

Fixes #5.