ci: update workflow to check document revision #8
This commit updates the `lint-commits.yml` workflow to ensure that the
document revision has been updated to the prospective version bump for
the changes within a PR.

This has been added to the `lint-commits.yml` workflow since that
workflow already runs when a PR is opened, and already generates a
prospective changelog and version for the changes.

Fixes #8.