fix(config): fix heading for new contributors #13
Previously, the heading for the "New Contributors" section in the
changelog was mistakenly a level 2 heading (`##`). This commit corrects
it to be on level 3, ensuring it is grouped under the release heading
which is level 2. This ensures that when the full changelog is
generated, the "New Contributors" section appears as a sub-section under
the release.

Fixes #13.