ci: ensure initial commits are conventional #10
This commit adds a new GitHub workflow to amend the initial commit of
repositories created from this template, to ensure that it conforms to
the conventional commits specification.

Previously, the default "Initial commit" message would be used, which
would cause the PR and build workflows fail until the initial commit was
manually rebased.

The workflow will delete itself while amending the initial commit,
as it will no longer be needed after the first commit is amended.

Fixes #10.