ci: skip build workflow on initial commit #16
This commit updates the `build-release.yml` workflow so that it skips
the initial commit by adding a new `eval` job to the workflow which
evaluates the workflow preconditions.

This is done since the initial commit will only contain the template
document, so is not a meaningful release. The workflow will also fail
for the initial commit, since the initial commit message won't
conform to the conventional commits specification until the
`create-from-template.yml` workflow runs.

Fixes #16.