ci: post preview document and changelog on PRs #7
This commit adds new functionality to the workflows to post a preview
build of the document, along with the proposed changelog, as a comment
on any opened PRs.

To facilitate this, I've moved the original `prep` stage from
`.github/workflows/build-release.yml` into a new reusable workflow;
`.github/workflows/prep-build.yml`. This allows it to be shared between
the release build and PR preview builds.

I've also updated the way the checkout steps work for building the
document and generating the changelog to try and ensure that they will
work for PRs opened from forks.

Fixes #7.